requirements.txt
SQLAlchemy==1.4.44
python-dotenv==1.0.0
pyTelegramBotAPI==4.12.0
openai==0.27.8
requests==2.31.0

